server/app-server/requirements.txt
numpy
six
pillow
requests
matplotlib
flask